Test put_bucket_logging with bucket policies
getmoto/moto#6715 enabled the use of put_bucket_logging when permissions
are granted on the logging bucket using a bucket policy.

When a target prefix is specified, and that prefix appears in the
Resource element of the statement granting PutObject permissions on the
bucket, put_bucket_logging fails with an InvalidTargetBucketForLogging
error, claiming the permissions are wrong.

The tests added here illustrate the problem.

I expect the unit test test_put_logging_w_bucket_policy_w_prefix to
pass. It fails.